Updated index.js and removed basicBot.js, added a basic README still needs work
